+++ browser/modules/state.js
@@ -71,6 +71,8 @@ var p, hashArr = hash.replace("#", "").split("/");
 // and hard code the call to reportRender.render in state::initializeFromHashPart
 let reportRender = require('../../extensions/conflictSearch/browser/reportRenderAlt');
 
+let activeLayersInSnapshot = false;
+
 /**
  * Returns internaly stored global state
  *
@@ -175,6 +177,15 @@ module.exports = {
                     legend.init().then(function () {
                         console.log("Vidi is now loaded");// Vidi is now fully loaded
                         window.status = "all_loaded";
+                        if (urlVars?.readyCallback) {
+                            try {
+                                window.parent.postMessage({
+                                    type: "vidiCallback",
+                                    method: urlVars.readyCallback
+                                }, "*");
+                            } catch (e) {
+                            }
+                        }
                     });
                 });
 
@@ -511,6 +522,7 @@ module.exports = {
                                 console.log("No active layers in snapshot");
                             } else {
                                 console.log("Active layers in snapshot");
+                                activeLayersInSnapshot = true;
                             }
                             this.applyState(state.snapshot).then(initResolve).catch((error) => {
                                 console.error(error)
@@ -796,5 +808,9 @@ module.exports = {
 
     resetStore: () => {
         _setInternalState({});
+    },
+
+    activeLayersInSnapshot: () => {
+        return activeLayersInSnapshot;
     }
 };