+++ extensions/conflictSearch/server/index.js
@@ -190,7 +190,7 @@ router.post('/api/extension/conflictSearch', function (req, response) {
                                 if (obj[hit].hits > 0) {
                                     let data = [];
                                     let name = obj[hit].title || obj[hit].table;
-                                    name = name.slice(0,30);
+                                    name = name.slice(0,30); // TODO also strip invalid characters
                                     if (names.includes(name)) {
                                         name = name.slice(0, -1) + postfixNumber;
                                         postfixNumber++;
@@ -212,7 +212,11 @@ router.post('/api/extension/conflictSearch', function (req, response) {
                                     }
                                     let ws = XLSX.utils.aoa_to_sheet(data);
                                     dataAdded = true;
-                                    XLSX.utils.book_append_sheet(wb, ws, name);
+                                    try {
+                                        XLSX.utils.book_append_sheet(wb, ws, name);
+                                    } catch (e) {
+                                        console.log(e.message);
+                                    }
                                 }
                             }
                         }