+++ CHANGELOG.md
@@ -4,12 +4,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
-## [UNRELEASED]
+## [2021.8.1] - 2021-25-8
+### Changed
+- Docker base image update to debian:bullseye-slim.
+
+### Added
+- Callback functions in embed.js. Two functions kan be defined in the parent page like this (with ``data-vidi-frame-name="map1"``):
+  - When Vidi is ready:
+  ```JavaScript
+    embedApi.vidiReady["map1"] = () => {
+        console.log("Vidi is ready")
+    }
+  ```
+  - When active layers from the snapshot is ready:
+  ```JavaScript
+    embedApi.activeLayersReady["map1"] = () => {
+        console.log("Active layers are ready")
+    }
+  ```
+
+### Fixed
+- In conflictSearch module, the adding of sheets to excel is now in a try/catch so invalid sheet names won't crash the Node process. 
+
+## [2021.8.0] - 2021-12-8
 ### Changed
 - `select_function` is now called when opening a panel in a accordion pop-up.
 
 ### Fixed
 - Print and Conflict modules will not longer keep unnecessary Meta data in state. This reduce the data volume. 
+- Layer tree will now build with all valid layers. Before it would start building as soon an invalid layer was proccess leaving out the rest.
 
 ## [2021.7.1] - 2021-8-7
 ### Fixed